Fix  Animate On Scroll : S1-PBI1 [Home Page] resolve #1 & Revisi+Debugging All

+++ Anjaman/resources/views/admin/transaksi.blade.php
@@ -57,7 +57,7 @@ Anjaman | Transaksi
                                                     <form action="/admin/deleteorder/{{$item->id}}" method="post">
                                                         @csrf
                                                         @method('delete')
-                                                        <button type="submit" href="/admin/deleteorder/{{$item->id}} " class="btn btn-danger">
+                                                        <button type="submit" href="/admin/deleteorder/{{$item->id}} " class="btn btn-danger" onclick="return confirm('Are You Sure?')">
                                                             <i class="fa fa-trash"></i>
                                                         </button>
                                                     </form>
